requirements.txt
opencv-python~=4.5.2.52
Pillow~=8.2.0